Inclusion of vkBasalt
Summary:
Fixes T8508. This should provide a Vulkan-only post processing layer akin to Reshade.
This is used best with a GOverlay - a GUI that help you configure it in a much easier way.

Test Plan:
Configure it using GOverlay.
Test CAS, FXAA and SMAA in a few different DXVK games on Intel and Nvidia GPU.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T8508

Differential Revision: https://dev.getsol.us/D9234